Added new event page
