+++ apps/web-app/templates/Base/index.tsx
@@ -2,7 +2,7 @@ import Header from "../../components/Header"
 import Footer from "../../components/Footer"
 
 const BaseTemplate = ({ children }: any) => (
-    <div className="flex flex-col relative">
+    <div className="flex flex-col">
         <header>
             <Header />
         </header>
